requirements.txt
transformers==4.18.0
pytorch-nlp==0.5.0
pylint==2.13.5